requirements.txt
python-dotenv==1.0.0
python-telegram-bot==20.6
requests==2.31.0
SQLAlchemy==2.0.22
